REFACTOR - custom hooks for PostDetailScreen
